+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-- Fix file each message without given block.
+- Fix file each message method and message iterator each method
+  without given block.
 
 ## CRuby and mruby version 0.3.0 - 2026-03-17